[Modify] Login, Splash 화면 토끼 이모지 수정
- RabbitFace 이미지 변경
- RabbitFace 컴포넌트 분리
